Project: finally chose a name for the OS!
Since I was inspired by Serenty OS to get back into hobby OS
development, it feel fitting to chose a name in a similar vain.

As such, this is now Journey OS. I decided on the name since for
this project I'm more interested in the journey of of learning
while building than ever having a piece of software that's
useful.

It's all about the learning and all the bugs we make along the
way.