Allocator: implement a simple linked list allocator
This allocator design is more or less copied from the the blog os
project.

It implements a linked list by storing data inside free regions
of the allocation heap. The data consists of a size and option
for a next block.

To free memory, a new data block is created in the memory that is to
be freed and it is inserted as the new "head" node of the list.

This design necessitates that each allocation is at least the size
and also alignment of such a node block, which adds most of the
complications in this implementation.

It also currently has no way of dealing with internal
fragmentation.