Allocator: add simple list merging mechanism
This algorithm is quite inefficient, but at least it's something
and I feel it can be fixed once it becomes a problem.

For every memory node in the heap, every other node will be checked
if it's a neighbouring node and if so, the two are merged.

Because that involves inserting the new region at the start and it's
not really known where in the list the region was, we have to start
the whole process from the top.

This runs until there are no more neighbours found for any node.